delete oparation success

+++ src/pages/MyPostedJob/MyPostedJob.jsx
@@ -2,6 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../context/AuthProvider/AuthProvider";
 import axios from "axios";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 const MyPostedJob = () => {
     const {user} = useContext(AuthContext);
     const [mypostedjob,setMyPostedJob] = useState(null);
@@ -13,10 +14,31 @@ const MyPostedJob = () => {
     },[user])
 
     const handleDelete = (_id) => {
-      axios.delete(`http://localhost:5000/jobdelete?id=${_id}`)
-      .then(result => {
-        console.log(result.data)
-      })
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.delete(`http://localhost:5000/jobdelete?id=${_id}`)
+          .then(result => {
+            if(result.data.deletedCount){
+              const filterDelete = mypostedjob.filter(job => job._id !== _id)
+              setMyPostedJob(filterDelete);
+            }
+          })
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your job has been deleted.",
+            icon: "success",
+          });
+        }
+      });
+      
     } 
     
     if(!mypostedjob){
@@ -27,39 +49,52 @@ const MyPostedJob = () => {
 
     return (
       <div className="relative overflow-x-auto py-14">
-        <h1 className="text-center text-3xl pb-10">{mypostedjob?.length > 0 ? 'Posted Job Avalable' : 'No Posted Jobs'}</h1>
-        <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-          <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
-            <tr>
-              <th scope="col" className="px-6 py-3">
-                Category
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Job Title
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Email
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Min-Salary
-              </th>
-              <th scope="col" className="px-6 py-3">
-                Max-Salary
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-           {
-            mypostedjob?.map(posted => {
+        <div className={mypostedjob?.length > 0 ? "" : "mt-40"}>
+          <h1 className="text-center text-4xl pb-10">
+            {mypostedjob?.length > 0 ? "Posted Job Avalable" : "No Posted Jobs"}
+          </h1>
+          {!mypostedjob.length > 0 ? (
+            <img
+              className="max-w-[200px] mx-auto"
+              src="https://cdn3.vectorstock.com/i/1000x1000/00/07/not-available-flat-icon-vector-12770007.jpg"
+              alt=""
+            />
+          ) : (
+            ""
+          )}
+        </div>
+
+        <div className={mypostedjob?.length > 0 ? "block" : "hidden"}>
+          <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
+            <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+              <tr>
+                <th scope="col" className="px-6 py-3">
+                  Category
+                </th>
+                <th scope="col" className="px-6 py-3">
+                  Job Title
+                </th>
+                <th scope="col" className="px-6 py-3">
+                  Email
+                </th>
+                <th scope="col" className="px-6 py-3">
+                  Min-Salary
+                </th>
+                <th scope="col" className="px-6 py-3">
+                  Max-Salary
+                </th>
+              </tr>
+            </thead>
+            <tbody>
+              {mypostedjob?.map((posted) => {
                 const {
                   job_title,
-                  
+
                   category,
                   max_salary,
                   min_salary,
-                    _id,
+                  _id,
                   email,
-                  
                 } = posted || {};
                 return (
                   <tr
@@ -85,16 +120,19 @@ const MyPostedJob = () => {
                       </Link>
                     </td>
                     <td className="px-6 py-4">
-                      <button onClick={() => handleDelete(_id)} className="bg-rose-500 text-white px-2 py-1 rounded-md">
+                      <button
+                        onClick={() => handleDelete(_id)}
+                        className="bg-rose-500 text-white px-2 py-1 rounded-md"
+                      >
                         Delete
                       </button>
                     </td>
                   </tr>
                 );
-            })
-           }
-          </tbody>
-        </table>
+              })}
+            </tbody>
+          </table>
+        </div>
       </div>
     );
 };
